pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "namicode-cli"
version = "0.0.14"
description = "Agentic Coding Tool that runs in your terminal"
authors = [{ name = "Babitdor", email = "babitdorbryan14@gmail.com" }]
readme = {file = "README.md", content-type = "text/markdown"}
license = { text = "MIT" }
keywords = ["nami", "code-cli", "nami-code", "agentic-code-cli"]

requires-python = ">=3.11,<4.0"
classifiers = [
     "Development Status :: 3 - Alpha",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
     "Framework :: LangChain-DeepAgents",
 ]
dependencies = [
    "requests",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.52",
    "langchain-openai>=0.1.0",
    "tavily-python",
    "python-dotenv",
    "daytona>=0.113.0",
    "modal>=0.65.0",
    "markdownify>=0.13.0",
    "langchain>=1.0.7",
    "runloop-api-client>=0.69.0",
    "langchain-ollama>=1.0.0",
    "langchain-google-genai>=3.2.0",
    "transformers>=4.57.3",
    "mcp>=1.0.0",
    "docker>=7.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "langsmith>=0.1.0",
    "keyring>=24.0.0",
    "nest-asyncio>=1.6.0",
    "pillow",
    "e2b-code-interpreter>=1.0.0",
    "textual>=1.0.0",
    "textual-autocomplete>=3.0.0",
    "aiosqlite>=0.19.0"
]

[project.urls]
    Homepage = "https://github.com/Babitdor/namicode"
    Repository = "https://github.com/Babitdor/namicode"
    Issues = "https://github.com/Babitdor/namicode/issues"
    Changelog = "https://github.com/Babitdor/Namicode/blob/main/CHANGELOG.md"

[project.scripts]
nami = "namicode_cli.main:cli_main"

[tool.ruff]
line-length = 100
exclude = []

[tool.ruff.format]
docstring-code-format = true  # Formats code blocks in docstrings

[tool.ruff.lint]
select = [
    "ALL"  # Enable all rules by default
]
ignore = [
    "COM812",  # Messes with the formatter
    "ISC001",  # Messes with the formatter
    "PERF203", # Rarely useful
    "SLF001",  # Private member access
    "PLC0415", # Imports should be at the top. Not always desirable
    "PLR0913", # Too many arguments in function definition
    "PLC0414", # Inconsistent with how type checkers expect to be notified of intentional re-exports
    "C901",    # Too complex
]
unfixable = ["B028"]  # Rules that shouldn't be auto-fixed

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"  # Google-style docstrings
ignore-var-parameters = true

[tool.ruff.lint.per-file-ignores]
"namicode_cli/cli.py" = [
    "T201",    # Allow print statements in CLI
]
"tests/*" = [
    "D1",      # Skip documentation rules in tests
    "S101",    # Allow asserts in tests
    "S311",    # Allow pseudo-random generators in tests
    "ANN201",  # Missing return type annotation
    "INP001",  # Implicit namespace package
    "PLR2004", # Magic value comparisons are fine in tests
]

[tool.pytest.ini_options]
timeout = 10  # Default timeout for all tests (can be overridden per-test)

[tool.mypy]
strict = true
ignore_missing_imports = true
enable_error_code = ["deprecated"]
# Optional: reduce strictness if needed
disallow_any_generics = false
warn_return_any = false

[tool.uv.sources]
nami-deepagents = { path = "./deepagents-nami" }